+++ dependency_injector/__init__.py
@@ -1,7 +1,7 @@
 """Dependency injector."""
 
 from .catalog import AbstractCatalog
-from .catalog import CatalogSubset
+from .catalog import CatalogBundle
 from .catalog import override
 
 from .providers import Provider
@@ -31,7 +31,8 @@ from .utils import is_kwarg_injection
 from .utils import is_attribute_injection
 from .utils import is_method_injection
 from .utils import is_catalog
-from .utils import is_catalog_subset
+from .utils import is_catalog_bundle
+from .utils import ensure_is_catalog_bundle
 
 from .errors import Error
 
@@ -39,7 +40,7 @@ from .errors import Error
 __all__ = (
     # Catalogs
     'AbstractCatalog',
-    'CatalogSubset',
+    'CatalogBundle',
     'override',
 
     # Providers
@@ -72,7 +73,8 @@ __all__ = (
     'is_attribute_injection',
     'is_method_injection',
     'is_catalog',
-    'is_catalog_subset',
+    'is_catalog_bundle',
+    'ensure_is_catalog_bundle',
 
     # Errors
     'Error',