+++ dependency_injector/providers.py
@@ -18,11 +18,12 @@ class Provider(object):
     """Base provider class."""
 
     __IS_PROVIDER__ = True
-    __slots__ = ('overridden_by',)
+    __slots__ = ('overridden_by', 'bind')
 
     def __init__(self):
         """Initializer."""
         self.overridden_by = None
+        self.bind = None
 
     def __call__(self, *args, **kwargs):
         """Return provided instance."""
@@ -72,6 +73,11 @@ class Provider(object):
     def reset_override(self):
         """Reset all overriding providers."""
         self.overridden_by = None
+
+    @property
+    def is_bound(self):
+        """Check if provider is bound to any catalog."""
+        return bool(self.bind)
 
 
 class Delegate(Provider):