+++ examples/catalogs/bundles/services.py
@@ -1,0 +1,17 @@
+"""Example services."""
+
+
+class BaseService(object):
+    """Example base class of service."""
+
+
+class UsersService(BaseService):
+    """Example users service."""
+
+
+class AuthService(BaseService):
+    """Example auth service."""
+
+
+class PhotosService(BaseService):
+    """Example photo service."""