+++ tests/test_catalog.py
@@ -71,52 +71,98 @@ class CatalogsInheritanceTests(unittest.TestCase):
                                   p32=CatalogC.p32))
 
 
-class CatalogSubsetTests(unittest.TestCase):
-    """Catalog subset test cases."""
+class CatalogProvidersBindingTests(unittest.TestCase):
+    """Catalog providers binding test cases."""
 
-    catalog = None
+    def test_provider_is_bound(self):
+        """Test that providers are bound to the catalogs."""
+        self.assertIs(CatalogA.p11.bind.catalog, CatalogA)
+        self.assertEquals(CatalogA.p11.bind.name, 'p11')
+
+        self.assertIs(CatalogA.p12.bind.catalog, CatalogA)
+        self.assertEquals(CatalogA.p12.bind.name, 'p12')
+
+    def test_provider_rebinding(self):
+        """Test that provider could not be bound twice."""
+        self.assertRaises(di.Error, type, 'TestCatalog', (di.AbstractCatalog,),
+                          dict(some_name=CatalogA.p11))
+
+
+class CatalogBundleTests(unittest.TestCase):
+    """Catalog bundle test cases."""
 
     def setUp(self):
         """Set test environment up."""
-        self.subset = CatalogC('p11', 'p12')
+        self.bundle = CatalogC.Bundle(CatalogC.p11,
+                                      CatalogC.p12)
 
-    def test_get_attr_from_subset(self):
-        """Test get providers (attribute) from subset."""
-        self.assertIs(self.subset.p11, CatalogC.p11)
-        self.assertIs(self.subset.p12, CatalogC.p12)
+    def test_get_attr_from_bundle(self):
+        """Test get providers (attribute) from catalog bundle."""
+        self.assertIs(self.bundle.p11, CatalogC.p11)
+        self.assertIs(self.bundle.p12, CatalogC.p12)
 
-    def test_get_attr_not_from_subset(self):
-        """Test get providers (attribute) that are not in subset."""
-        self.assertRaises(di.Error, getattr, self.subset, 'p21')
-        self.assertRaises(di.Error, getattr, self.subset, 'p22')
-        self.assertRaises(di.Error, getattr, self.subset, 'p31')
-        self.assertRaises(di.Error, getattr, self.subset, 'p32')
+    def test_get_attr_not_from_bundle(self):
+        """Test get providers (attribute) that are not in bundle."""
+        self.assertRaises(di.Error, getattr, self.bundle, 'p21')
+        self.assertRaises(di.Error, getattr, self.bundle, 'p22')
+        self.assertRaises(di.Error, getattr, self.bundle, 'p31')
+        self.assertRaises(di.Error, getattr, self.bundle, 'p32')
 
-    def test_get_method_from_subset(self):
-        """Test get providers (get() method) from subset."""
-        self.assertIs(self.subset.get('p11'), CatalogC.p11)
-        self.assertIs(self.subset.get('p12'), CatalogC.p12)
+    def test_get_method_from_bundle(self):
+        """Test get providers (get() method) from bundle."""
+        self.assertIs(self.bundle.get('p11'), CatalogC.p11)
+        self.assertIs(self.bundle.get('p12'), CatalogC.p12)
 
-    def test_get_method_not_from_subset(self):
-        """Test get providers (get() method) that are not in subset."""
-        self.assertRaises(di.Error, self.subset.get, 'p21')
-        self.assertRaises(di.Error, self.subset.get, 'p22')
-        self.assertRaises(di.Error, self.subset.get, 'p31')
-        self.assertRaises(di.Error, self.subset.get, 'p32')
+    def test_get_method_not_from_bundle(self):
+        """Test get providers (get() method) that are not in bundle."""
+        self.assertRaises(di.Error, self.bundle.get, 'p21')
+        self.assertRaises(di.Error, self.bundle.get, 'p22')
+        self.assertRaises(di.Error, self.bundle.get, 'p31')
+        self.assertRaises(di.Error, self.bundle.get, 'p32')
 
     def test_has(self):
-        """Test checks of providers availability in subsets."""
-        self.assertTrue(self.subset.has('p11'))
-        self.assertTrue(self.subset.has('p12'))
+        """Test checks of providers availability in bundle."""
+        self.assertTrue(self.bundle.has('p11'))
+        self.assertTrue(self.bundle.has('p12'))
 
-        self.assertFalse(self.subset.has('p21'))
-        self.assertFalse(self.subset.has('p22'))
-        self.assertFalse(self.subset.has('p31'))
-        self.assertFalse(self.subset.has('p32'))
+        self.assertFalse(self.bundle.has('p21'))
+        self.assertFalse(self.bundle.has('p22'))
+        self.assertFalse(self.bundle.has('p31'))
+        self.assertFalse(self.bundle.has('p32'))
 
-    def test_creating_with_undefined_provider(self):
-        """Test subset creation with provider that is not in catalog."""
-        self.assertRaises(di.Error, CatalogC, 'undefined_provider')
+    def test_create_bundle_with_unbound_provider(self):
+        """Test that bundle is not created with unbound provider."""
+        self.assertRaises(di.Error, CatalogC.Bundle, di.Provider())
+
+    def test_create_bundle_with_another_catalog_provider(self):
+        """Test that bundle can not contain another catalog's provider."""
+        class TestCatalog(di.AbstractCatalog):
+            """Test catalog."""
+
+            provider = di.Provider()
+
+        self.assertRaises(di.Error,
+                          CatalogC.Bundle, CatalogC.p31, TestCatalog.provider)
+
+    def test_create_bundle_with_another_catalog_provider_with_same_name(self):
+        """Test that bundle can not contain another catalog's provider."""
+        class TestCatalog(di.AbstractCatalog):
+            """Test catalog."""
+
+            p31 = di.Provider()
+
+        self.assertRaises(di.Error,
+                          CatalogC.Bundle, CatalogC.p31, TestCatalog.p31)
+
+    def test_is_bundle_owner(self):
+        """Test that catalog bundle is owned by catalog."""
+        self.assertTrue(CatalogC.is_bundle_owner(self.bundle))
+        self.assertFalse(CatalogB.is_bundle_owner(self.bundle))
+        self.assertFalse(CatalogA.is_bundle_owner(self.bundle))
+
+    def test_is_bundle_owner_with_not_bundle_instance(self):
+        """Test that check of bundle ownership raises error with not bundle."""
+        self.assertRaises(di.Error, CatalogC.is_bundle_owner, object())
 
 
 class CatalogTests(unittest.TestCase):
@@ -136,7 +182,7 @@ class CatalogTests(unittest.TestCase):
         self.assertRaises(di.Error, CatalogC.get, 'undefined')
 
     def test_has(self):
-        """Test checks of providers availability in subsets."""
+        """Test checks of providers availability in catalog."""
         self.assertTrue(CatalogC.has('p11'))
         self.assertTrue(CatalogC.has('p12'))
         self.assertTrue(CatalogC.has('p21'))
@@ -144,14 +190,6 @@ class CatalogTests(unittest.TestCase):
         self.assertTrue(CatalogC.has('p31'))
         self.assertTrue(CatalogC.has('p32'))
         self.assertFalse(CatalogC.has('undefined'))
-
-    def test_is_subset_owner(self):
-        """Test that catalog is subset owner."""
-        subset = CatalogA()
-
-        self.assertTrue(CatalogA.is_subset_owner(subset))
-        self.assertFalse(CatalogB.is_subset_owner(subset))
-        self.assertFalse(CatalogC.is_subset_owner(subset))
 
     def test_filter_all_providers_by_type(self):
         """Test getting of all catalog providers of specific type."""